+++ grad_roze/lib/widgets/Bouquet/BouquetViewModel.dart
@@ -3,7 +3,7 @@ import 'package:grad_roze/config.dart';
 class BouquetViewModel {
   final String id;
   final String name;
-  final int price;
+  final double price;
   final String businessName;
   final String imageUrl;
 
@@ -20,7 +20,7 @@ class BouquetViewModel {
     return BouquetViewModel(
         id: json['_id'],
         name: json['name'],
-        price: json['price'].toInt(),
+        price: json['price'].toDouble(),
         businessName: json['business'],
         imageUrl: '$url${json['imageURL']}');
   }
